INTERLOK-4081 Upgrade to jetty 10.
Wiremock is no longer working with jetty 10 so we use mockserver instead with a bit of twicking.
Request payload in the __files dir are not working anymore.
The mappings json format in the mappings dir are a bit different: request -> httpRequest, url -> path, response -> httpResponse, headers format changed, bodyFilename -> body in a json format.
USe lombok in WireMockHelper